chore: rename lib/theme-chalk to lib/style
