+++ lib/button/Button.vue
@@ -74,7 +74,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '../theme-chalk/common.scss';
+@import '../style/common.scss';
 
 $h: 32px;
 .coast-button {
