+++ lib/button/ButtonGroup.vue
@@ -22,7 +22,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '../theme-chalk/common.scss';
+@import '../style/common.scss';
 
 .coast-button-group {
   display: inline-flex;
