+++ lib/button/__tests__/button.spec.ts
@@ -3,7 +3,7 @@ import { ref } from 'vue';
 import Button from '../Button.vue';
 import ButtonGroup from '../ButtonGroup.vue';
 
-jest.mock('../../theme-chalk/iconfont/index.js');
+jest.mock('../../style/iconfont/index.js');
 
 const AXIOM = 'Tomorrow will be even better';
