+++ lib/card/Card.vue
@@ -81,7 +81,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '../theme-chalk/common.scss';
+@import '../style/common.scss';
 
 $border-color: #eaeaea;
 .coast-card {
