+++ lib/input/Input.vue
@@ -185,7 +185,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '../theme-chalk/common.scss';
+@import '../style/common.scss';
 
 $fontSize: 14px;
 $height: 36px;
