+++ lib/input/__tests__/input.spec.ts
@@ -3,7 +3,7 @@ import { mount } from '@vue/test-utils';
 import Input from '../Input.vue';
 import Icon from '../../icon/Icon.vue';
 
-jest.mock('../../theme-chalk/iconfont/index.js');
+jest.mock('../../style/iconfont/index.js');
 
 const AXIOM = 'Tomorrow will be even better';
