+++ lib/switch/Switch.vue
@@ -26,7 +26,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '../theme-chalk/common.scss';
+@import '../style/common.scss';
 
 $h: 22px;
 $h2: $h - 4px;
