+++ lib/toast/style/toast.scss
@@ -1,4 +1,4 @@
-@import '../../theme-chalk/common.scss';
+@import '../../style/common.scss';
 
 .coast-toast {
   position: fixed;
